Abstract some Splitter handler functionality to a base trait
also add in Transformer#collect to mirror Transformer.Collect from
the previous commit
